Adding Bear textures
The bear material is added, the it doesnt work with the animated model right now because of the UV issue but works fine on a stationary model. I added it to the sample scene to test, if adding it to any bear model just make sure to add it to both eye objects too.